+++ changelog.md
@@ -12,3 +12,8 @@
 ## v1.1.0
 - added custom styling to drivetrain buttons again
 - corrected icons
+
+## v1.1.1
+- fix onInputChange by using `onChange`
+- fix Autocomplete value not clearing on submit by using `value`
+- fix rendering ternary in db-view